refactor: federated issuer validation
